NXP-23417: Fix zip directory as a jar for Windows
